+++ CHANGELOG.md
@@ -46,6 +46,7 @@ Version 1.5.1 (not released yet)
 - fix shooting through transparent sections of DDS textures
 - change default server update rate to 20
 - add 'update_rate' command server-side and client-side
+- fix object interpolation in multi-player being played too fast causing a possible jitter
 
 Version 1.5.0 (released 2020-09-05)
 -----------------------------------